+++ src/Tests/WB.Tests.Unit/BoundedContexts/Supervisor/Synchronization/InterviewsSynchronizerTests/Push/when_pushing_interview_and_interview_has_one_event_with_empty_origin.cs
@@ -51,7 +51,7 @@ namespace WB.Tests.Unit.BoundedContexts.Supervisor.Synchronization.InterviewsSyn
 
             interviewEvent = Create.CommittedEvent(eventSourceId: interviewId, origin: null);
             var eventStore = Mock.Of<IEventStore>(store
-                => store.ReadFrom(interviewId, 0, Moq.It.IsAny<long>()) == new CommittedEventStream(interviewId, interviewEvent));
+                => store.ReadFrom(interviewId, 0, Moq.It.IsAny<int>()) == new CommittedEventStream(interviewId, interviewEvent));
 
             var interviewSummaryRepositoryWriter = Mock.Of<IReadSideRepositoryReader<InterviewSummary>>(writer
                 => writer.GetById(interviewId.FormatGuid()) == Create.InterviewSummary());
